# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Linkbit Controller API
  version: 0.1.0
servers:
  - url: https://controller.example.com
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Linkbit-API-Key
    DeviceTokenAuth:
      type: apiKey
      in: header
      name: X-Linkbit-Device-Token
paths:
...
  /api/v1/relays/{id}:
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /api/v1/devices/{id}/network-config:
    get:
      security:
        - DeviceTokenAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device network config
...